backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pgvector==0.3.6
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
httpx==0.28.1
numpy==2.2.1
PyMuPDF==1.25.3
resend==2.7.0
rq==2.1.0
redis==5.2.1
openai==1.59.5
python-multipart==0.0.20
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-httpx==0.35.0
aiosqlite==0.20.0
greenlet==3.1.1
